Uploaded initial version
Designed the structure of the game based off of the Pong lesson. Currently, the ship can move left and right without going off screen. A projectile is shot from the ship by pressing space but I am working on figuring out how to shoot multiple. Enemies are stored in a vector and generated in a loop.