Added interactions between the enemies and projectiles
Right now, the projectiles and enemies are being stored in vectors
When the space bar is pressed a new projectile is created and added to the vector
There is a nested for loop checking the projectiles and enemies vectors for a collision, which is slowing down the game majorly. I am going to try to combat this with a limited number of projectiles to be created at a time.